refactor: take out RadioColumn, simplify EdgeView
